Added more detail across the website
Updated project descriptions to be more technical and explanatory rather than short and sweet.
Updated my philosophy, journey, and learning opportunities to be much stronger and more professional.
Turned the card elements in both index and about to be links to the website with full mobile compatibility.
Added "What Am I Looking For" in about.
Fixed the titles for X (Formerly Twitter) and Open Graph / Facebook.
Updated tag selection on the projects page to allow the user to click on the tag in the project card to set the master tag.
Updated the date element to be in en-GB instead of en-US.